refactor and add breacrumbs
