+++ src/models/Events/EventAdmin.ts
@@ -1,0 +1,6 @@
+export interface EventAdmin {
+  userId: string;
+  fullName: string;
+  adminType: string;
+  avatarUrl: string;
+}
