+++ src/models/Events/EventDetails.ts
@@ -1,0 +1,13 @@
+import { EventInformation } from "./EventInformation";
+
+export interface EventDetails {
+  event: EventInformation;
+  participantAssessment: number;
+  isUserParticipant: boolean;
+  isUserApprovedParticipant: boolean;
+  isUserUndeterminedParticipant: boolean;
+  isUserRejectedParticipant: boolean;
+  isEventFinished: boolean;
+  isEventNotApproved: boolean;
+  canEstimate: boolean;
+}
