+++ src/models/Events/EventGallery.ts
@@ -1,0 +1,5 @@
+export interface EventGallery {
+  galleryId: number;
+  fileName: string;
+  encodedData: string;
+}
