+++ src/models/Events/EventParticipant.ts
@@ -1,0 +1,9 @@
+export interface EventParticipant {
+  participantId: number;
+  fullName: string;
+  email: string;
+  userId: string;
+  statusId: number;
+  status: string;
+  wasPresent: boolean;
+}
