+++ src/pages/Actions/ActionEvent/EventCreate/EventCreate.module.css
@@ -1,92 +1,90 @@
 .background {
-    background-color: rgb(238, 238, 238);
+  background-color: rgb(238, 238, 238);
 }
 
 form {
-    width: 100%;
+  width: 100%;
 }
 
 .label {
-    background-color: #3c5438;
-    color: #f0f8ff;
-    margin: 17px;
-    padding: 8px;
-    text-align: center;
-    width: 150px;
+  background-color: #3c5438;
+  color: #f0f8ff;
+  margin: 17px;
+  padding: 8px;
+  text-align: center;
+  width: 150px;
 }
 
-[type='radio'] {
-    margin-right: 20px;
+[type="radio"] {
+  margin-right: 20px;
 }
 
 .select:hover,
 .select:focus,
 input:hover,
 .select1:hover {
-    outline: 0;
-    border-color: #3c5438 !important;
+  outline: 0;
+  border-color: #3c5438 !important;
 }
 
 .input {
-    text-align: start;
-    padding: 5px 11px;
-    font-size: 14px;
-    width: 100% !important;
-    
+  text-align: start;
+  padding: 5px 11px;
+  font-size: 14px;
+  width: 100% !important;
 }
 .formItem {
-    display: block;
-    width: 100% !important;
+  display: block;
+  width: 370px !important;
 }
 
 .select {
-    min-width: 370px;
-    max-width: 370px;
-    position: relative;
+  min-width: 370px;
+  max-width: 370px;
+  position: relative;
 }
 
 .button {
-    margin: 10px 0px 25px 7px;
-    border: none;
-    width: auto;
-    height: 40px;
-    outline: none;
-    cursor: pointer;
-    margin-bottom: 15px;
-    box-shadow: 0 7px 14px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
-    transition: all 0.3s cubic-bezier(.25, .8, .25, 1);
-    display: inline-table;
-    float: right;
+  margin: 10px 0px 25px 7px;
+  border: none;
+  width: auto;
+  height: 40px;
+  outline: none;
+  cursor: pointer;
+  margin-bottom: 15px;
+  box-shadow: 0 7px 14px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
+  transition: all 0.3s cubic-bezier(0.25, 0.8, 0.25, 1);
+  display: inline-table;
+  float: right;
 }
 
 .button:hover {
-    border: none;
-    outline: none;
-    box-shadow: 0 2px 0 rgba(0, 0, 0, 0.015);
-    transition: all 0.3s cubic-bezier(0.645, 0.045, 0.355, 1);
-    box-shadow: 0 14px 28px rgba(0, 0, 0, 0.25), 0 10px 10px rgba(0, 0, 0, 0.22);
+  border: none;
+  outline: none;
+  box-shadow: 0 2px 0 rgba(0, 0, 0, 0.015);
+  transition: all 0.3s cubic-bezier(0.645, 0.045, 0.355, 1);
+  box-shadow: 0 14px 28px rgba(0, 0, 0, 0.25), 0 10px 10px rgba(0, 0, 0, 0.22);
 }
 
 .radio {
-    width: 100%;
-    justify-content: center;
-    display: block;
+  width: 100%;
+  justify-content: center;
+  display: block;
 }
 
-@media(max-width:767px) {
-    .input {
-        width: 100% !important;
-    }
-    .select {
-        width: 100% !important;
-    }
-    .button {
-        margin: auto;
-        display: inline-block;
-        
-    }
-    .ant-radio-group {
-        display: flex;
-        flex-direction: column;
-    }
+@media (max-width: 767px) {
+  .input {
+    width: 100% !important;
+  }
+  .select {
+    width: 100% !important;
+  }
+  .button {
+    margin: auto;
+    display: inline-block;
+  }
+  .ant-radio-group {
+    display: flex;
+    flex-direction: column;
+  }
 }
