+++ src/pages/Actions/ActionEvent/EventInfo/EventInfo.module.css
@@ -1,15 +1,9 @@
 .userTableTitle {
-    color: #3c5438;
+  color: #3c5438;
 }
 
 .searchArea {
-    margin-bottom: 10px;
-    width: 30%;
-    align-self: center;
-}
-
-@media (max-width: 1000px) {
-    .searchArea {
-        width: 40%;
-    }
+  margin-bottom: 10px;
+  max-width: 500px;
+  align-self: center;
 }
