+++ src/pages/Actions/Actions.module.css
@@ -1,21 +1,17 @@
-.background{
-    background-color: rgb(238, 238, 238);
+.mainTitle {
+  text-align: center;
+  font-size: 40px;
+  color: #3c5438;
 }
 
-.mainTitle{
-    text-align: center;
-    font-size: 40px;
-    color:#3c5438;
+.actionsWrapper {
+  display: flex;
+  justify-content: center;
+  flex-wrap: wrap;
+  width: 90%;
+  margin: 0 auto;
 }
 
-.actionsWrapper{
-    display: flex;
-    justify-content: center;
-    flex-wrap: wrap;
-    width: 90%;
-    margin: 0 auto;
+.pagination {
+  text-align: center;
 }
-
-.pagination{
-    text-align: center;
-  }
