+++ src/pages/Actions/EventTypes/EventTypes.module.css
@@ -1,6 +1,3 @@
-.background {
-  background-color: rgb(238, 238, 238);
-}
 .mainTitle {
   text-align: center;
   font-size: 40px;
